[IMP] web: removed low opacity on checkboxes when not in edit mode
It didn't work in other browsers than Firefox and it actually was an issue as it made them hard to read.
The disabled state is enough to tell that they are not editable, we don't need to hinder the readability.
Maybe this css rule come from a time where the disabled state was not applied ?
Fixes opw 630611